fix relative positioned block link click issue
To make square background benethe the heading we added 'position:
relative;' with z-index: -1; and to note cause this square background
go benethe its container we added its continer also position: relative
and z-index: -2; but this caused the link in the container unclickable.
 We fixed this by removing all these relative positions instead we
added relative position to the next block which right below the square
background which is the content wrapper.